% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_balance_sheet.R
\name{GetBalanceSheet}
\alias{GetBalanceSheet}
\title{Acquire balance sheet.}
\usage{
GetBalanceSheet(symbol, year, quarter)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company of interest}

\item{year}{A numeric vector specifying the year during which the report was filed}

\item{quarter}{A character vector specifying the quarter during which the report was filed}
}
\value{
A data frame representing a company's quarterly balance sheet from a given filing year
}
\description{
Extracts and displays balance sheet from the quarterly report of a given company.
}
\examples{
\dontrun{
GetBalanceSheet("FB", 2016, "Q1")
}
}
