% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cash_flow.R
\name{GetCashFlow}
\alias{GetCashFlow}
\title{Acquire statement of cash flow.}
\usage{
GetCashFlow(symbol, year, quarter)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company of interest}

\item{year}{A numeric vector specifying the year during which the report was filed}

\item{quarter}{A character vector specifying the quarter during which the report was filed}
}
\value{
A data frame representing a company's quarterly cash flow statement from a given filing year
}
\description{
Extracts and displays statement of cash flow from the quarterly report of a given company.
}
\examples{
\dontrun{
GetCashFlow("FB", 2016, "Q1")
}
}
