% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_income.R
\name{GetIncome}
\alias{GetIncome}
\title{Acquire income statement.}
\usage{
GetIncome(symbol, year, quarter)
}
\arguments{
\item{symbol}{A character vector specifying the stock symbol of the company of interest}

\item{year}{A numeric vector specifying the year during which the report was filed}

\item{quarter}{A character vector specifying the quarter during which the report was filed}
}
\value{
A data frame representing a company's quarterly income statement from a given filing year
}
\description{
Extracts and displays income statement from the quarterly report of a given company.
}
\examples{
\dontrun{
GetIncome("FB", 2016, "Q1")
}
}
